don't show creation links in diffs
In confluence, when you link a non-existing page, a link that creates
that page is inserted. For whatever reason, an ID is contained in that
link, and everytime somebody changes something in the page, every
creation link gets a new ID. This shows up in the diff but is totally
uninteresting. So this commit filters that by removing the ID from the
href-attribute of creation links.